refactor: move package registry files
