+++ packages/api/src/package-registry/jsr/__tests__/ext-pkg.test.ts
